+++ src/main/java/pl/asie/computronics/integration/waila/providers/WailaLocomotiveRelay.java
@@ -4,6 +4,7 @@ import mcp.mobius.waila.api.IWailaConfigHandler;
 import mcp.mobius.waila.api.IWailaDataAccessor;
 import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NBTTagCompound;
+import pl.asie.computronics.integration.waila.ConfigValues;
 import pl.asie.computronics.util.StringUtil;
 
 import java.util.List;
@@ -17,9 +18,9 @@ public class WailaLocomotiveRelay extends ComputronicsWailaProvider {
 	public List<String> getWailaBody(ItemStack itemStack, List<String> currenttip, IWailaDataAccessor accessor,
 		IWailaConfigHandler config) {
 
-		/*if(!ConfigValues.RelayBound.getConfig(config)) {
+		if(!ConfigValues.RelayBound.getValue(config)) {
 			return currenttip;
-		}*/
+		}
 
 		NBTTagCompound nbt = accessor.getNBTData();
 		String boundKey = "tooltip.computronics.waila.relay." + (nbt.getBoolean("bound") ? "bound" : "notbound");