+++ src/main/java/pl/asie/computronics/integration/waila/providers/WailaPeripheral.java
@@ -6,6 +6,7 @@ import mcp.mobius.waila.api.IWailaConfigHandler;
 import mcp.mobius.waila.api.IWailaDataAccessor;
 import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NBTTagCompound;
+import pl.asie.computronics.integration.waila.ConfigValues;
 import pl.asie.computronics.reference.Mods;
 import pl.asie.computronics.util.StringUtil;
 
@@ -20,9 +21,9 @@ public class WailaPeripheral extends ComputronicsWailaProvider {
 	public List<String> getWailaBody(ItemStack itemStack, List<String> currenttip, IWailaDataAccessor accessor,
 		IWailaConfigHandler config) {
 
-		/*if(!ConfigValues.Address.getConfig(config)) {
+		if(!ConfigValues.Address.getValue(config)) {
 			return currenttip;
-		}*/
+		}
 
 		NBTTagCompound nbt = accessor.getNBTData();
 		if(Loader.isModLoaded(Mods.OpenComputers)) {